refactor: 뉴스스탠드 헤더 CLASS 적용
