--- src/components/news-stand-header/main.js
@@ -1,8 +1,0 @@
-import { NewsStandHeaderSubject } from './subject.js';
-import { NewsStandHeaderObserver } from './observer.js';
-
-const newsStandHeaderSubject = new NewsStandHeaderSubject();
-const newsStandHeaderObserver = new NewsStandHeaderObserver(newsStandHeaderSubject);
-newsStandHeaderSubject.addObserver(newsStandHeaderObserver);
-
-export default newsStandHeaderSubject;
