--- src/components/news-stand-header/observer.js
@@ -1,14 +1,0 @@
-import { Observer } from '../../utils/observer.js';
-
-class NewsStandHeaderObserver extends Observer {
-  constructor(subject) {
-    super(subject);
-  }
-
-  update() {
-    location.reload();
-    this.subject.updateHeaderTime();
-  }
-}
-
-export { NewsStandHeaderObserver };
